# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OptiPlan360 API",
    "description": "Comprehensive API documentation for OptiPlan360 - Integrated Woodworking Management System",
    "version": "1.0.0",
    "contact": {
      "name": "OptiPlan360 Development Team",
      "email": "dev@optiplan360.com",
      "url": "https://optiplan360.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.optiplan360.com/v1",
      "description": "Production Server"
    },
    {
      "url": "https://staging-api.optiplan360.com/v1",
      "description": "Staging Server"
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get Orders List",
        "description": "Retrieve paginated list of orders with optional filtering",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PENDING_APPROVAL", "IN_PRODUCTION", "COMPLETED", "DELIVERED", "CANCELLED"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create New Order",
        "description": "Create a new order with validation",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get Order Details",
        "description": "Retrieve detailed information about a specific order",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "tags": ["Orders"],
        "summary": "Update Order",
        "description": "Update existing order information",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete Order",
        "description": "Delete an existing order",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
    "/ocr/upload": {
      "post": {
        "tags": ["OCR"],
        "summary": "Upload Document for OCR Processing",
        "description": "Upload a document and process it with OCR",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document file to process"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OCR processing initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateOrderRequest": {
        "type": "object",
        "required": ["customer_id", "product_id", "quantity", "thickness"],
        "properties": {
          "customer_id": {
            "type": "integer",
            "example": 1
          },
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "quantity": {
            "type": "number",
            "minimum": 0.001,
            "example": 100
          },
          "thickness": {
            "type": "number",
            "minimum": 1,
            "maximum": 50,
            "example": 18
          },
          "width": {
            "type": "number",
            "minimum": 1,
            "example": 2100
          },
          "height": {
            "type": "number",
            "minimum": 1,
            "example": 2970
          },
          "notes": {
            "type": "string",
            "example": "Special requirements"
          }
        }
      },
      "UpdateOrderRequest": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "quantity": {
            "type": "number"
          },
          "thickness": {
            "type": "number"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PENDING_APPROVAL", "IN_PRODUCTION", "COMPLETED", "DELIVERED", "CANCELLED"]
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "order_number": {
            "type": "string",
            "example": "SPR20260314001"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "quantity": {
            "type": "number",
            "example": 100
          },
          "thickness": {
            "type": "number",
            "example": 18
          },
          "width": {
            "type": "number",
            "example": 2100
          },
          "height": {
            "type": "number",
            "example": 2970
          },
          "status": {
            "type": "string",
            "enum": ["DRAFT", "PENDING_APPROVAL", "IN_PRODUCTION", "COMPLETED", "DELIVERED", "CANCELLED"],
            "example": "DRAFT"
          },
          "total_area": {
            "type": "number",
            "example": 6.237
          },
          "notes": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          }
        }
      },
      "OrdersResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Test Customer"
          },
          "phone": {
            "type": "string",
            "example": "+905551234567"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "customer@example.com"
          },
          "address": {
            "type": "string",
            "example": "Istanbul, Turkey"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-14T10:30:00Z"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "ts_code": {
            "type": "string",
            "example": "OAK001"
          },
          "name": {
            "type": "string",
            "example": "Oak Wood 18mm"
          },
          "material_type": {
            "type": "string",
            "example": "Oak"
          },
          "thickness": {
            "type": "number",
            "example": 18
          },
          "price_per_m2": {
            "type": "number",
            "example": 150.50
          },
          "in_stock": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "total": {
            "type": "integer",
            "example": 125
          },
          "total_pages": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "OCRResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "example": "ocr_task_123456"
          },
          "status": {
            "type": "string",
            "enum": ["processing", "completed", "failed", "requires_review"],
            "example": "processing"
          },
          "confidence_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 85.5
          },
          "extracted_data": {
            "type": "object",
            "example": {
              "customer_name": "Test Customer",
              "product_code": "OAK001",
              "quantity": 100
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed"
          },
          "message": {
            "type": "string",
            "example": "Required field missing: customer_id"
          },
          "details": {
            "type": "object",
            "example": {
              "field": "customer_id",
              "reason": "required"
            }
          }
        }
      }
    }
  },
...
}